+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # LessCSS Maven Plugin Changelog
 
+## 1.2.0
+Dependencies:
+* Upgrade [Extended LessCSS Compiler](http://lesscss-extended-compiler.projects.gabrys.biz/) from [2.0](http://lesscss-extended-compiler.projects.gabrys.biz/2.0/) to [2.1.0](http://lesscss-extended-compiler.projects.gabrys.biz/2.1.0/) (support for `classpath://` protocol)
+
+[See documentation](http://lesscss-maven-plugin.projects.gabrys.biz/1.2.0/)
+
 ## 1.1
 Dependencies:
 * Upgrade [Extended LessCSS Compiler](http://lesscss-extended-compiler.projects.gabrys.biz/) from [1.0](http://lesscss-extended-compiler.projects.gabrys.biz/1.0/) to [2.0](http://lesscss-extended-compiler.projects.gabrys.biz/2.0/)